+++ packages/react-i18n/CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ---
 
-<!-- ## [Unreleased] - -->
+## [Unreleased] -
+
+### Added
+
+- Add `unformatCurrency` utility return the normalized value of a formatted money value
 
 ## [0.7.4] - 2019-01-21